+++ NEWS.md
@@ -1,3 +1,3 @@
-# alpha-correction-bh 0.0.1
+# alpha-correction-bh 0.0.2
 
-* Added a `NEWS.md` file to track changes to the package.
+* Bugfix: All alphas should be non-significant once the first alpha value check comes back non-significant.